Add unidling support.
When all pods for a service VIP are removed, this used to result in the
load balancer for the VIP being removed from OVN.

With this change, we instead leave the VIP configured, but have the
backends in the OVN load balancer empty.

We also enable the controller_event northbound option. Now, with empty
load balancer backends, if OVN receives traffic for a load balancer VIP,
it will raise an empty_lb_backends event in the southbound database.

This patch adds an EventChecker method for the Controller that polls the
southbound database every 5 seconds to find if there are any unhandled
empty_lb_backend events. If there are any, then the event is parsed and
we send an event to kubernetes to request pods.

The use case for this is OpenShift's unidling. OpenShift will
automatically spin up pods if traffic is received on idled services.
This patchset allows for this behavior to work when using ovn-kubernetes
with OpenShift.