Add ACL reject rule for non-functional load balancers.
When endpoints are removed, then traffic sent to the load balancer VIP
should hit an ACL with a "reject" action.

When an endpoint is removed, then when the corresponding load balancer's
VIP is removed, then we find all logical switches that use that load
balancer. We then create an ACL to reject traffic to that VIP and add
the ACL to all of the retrieved logical switches.

When the endpoint is re-added, then the ACL is removed from all logical
switches.